requirements.txt
flask==2.2.5
flask-compress
numpy==1.23.5
pandas==2.0.0
tensorflow-cpu
librosa==0.10.1
pillow==9.5.0
pydub==0.24.1
matplotlib==3.7.1
static-ffmpeg
werkzeug
gunicorn